src/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-evaluation-sandbox"
version = "0.1.0"
description = "Agentic Evaluation Sandbox CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0.1",
]

[project.scripts]
aes = "aes.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "PyYAML>=6.0.1",
]

[tool.setuptools.packages.find]
where = ["src"]
